+++ Source/SwiftLintFramework/Rules/Performance/EmptyCollectionLiteralRule.swift
@@ -2,6 +2,7 @@ import SourceKittenFramework
 
 public struct EmptyCollectionLiteralRule: ConfigurationProviderRule, OptInRule, AutomaticTestableRule {
     public var configuration = SeverityConfiguration(.warning)
+    public var initializedWithNonEmptyConfiguration: Bool = false
 
     public init() {}