+++ Source/swiftlint/Extensions/Configuration+CommandLine.swift
@@ -231,7 +231,7 @@ extension Configuration {
 
     init(options: LintOrAnalyzeOptions) {
         let cachePath = options.cachePath.isEmpty ? nil : options.cachePath
-        self.init(subConfigQueue: [options.configurationFile], rootPath: type(of: self).rootPath(from: options.paths),
+        self.init(childConfigQueue: [options.configurationFile], rootPath: type(of: self).rootPath(from: options.paths),
                   optional: isConfigOptional(), quiet: options.quiet, enableAllRules: options.enableAllRules,
                   cachePath: cachePath)
     }
@@ -248,13 +248,13 @@ extension Configuration {
 
     init(options: AutoCorrectOptions) {
         let cachePath = options.cachePath.isEmpty ? nil : options.cachePath
-        self.init(subConfigQueue: [options.configurationFile], rootPath: type(of: self).rootPath(from: options.paths),
+        self.init(childConfigQueue: [options.configurationFile], rootPath: type(of: self).rootPath(from: options.paths),
                   optional: isConfigOptional(), quiet: options.quiet, cachePath: cachePath)
     }
 
     // MARK: Rules command
     init(options: RulesOptions) {
-        self.init(subConfigQueue: [options.configurationFile], rootPath: nil, optional: isConfigOptional())
+        self.init(childConfigQueue: [options.configurationFile], rootPath: nil, optional: isConfigOptional())
     }
 }