+++ CHANGELOG.md
@@ -220,6 +220,10 @@
 
 * Make `contains_over_first_not_nil` rule also match `first(where:) == nil`.  
   [Colton Schlosser](https://github.com/cltnschlosser)
+  
+* Add option to specify a `sub_config` file in any SwiftLint configuration file. Recursive sub configs are also possible.  
+  [Frederick Pietschmann](https://github.com/fredpi)
+  [#1352](https://github.com/realm/SwiftLint/issues/1352)
 
 * Add two new cases to the Mark rule to detect a Mark using three slashes.  
   [nvanfleet](https://github.com/nvanfleet)
@@ -367,6 +371,7 @@ This is the last release to support building with Swift 4.2.x.
 * Added 'file_name_no_space' opt-in rule.  
   [timcmiller](https://github.com/timcmiller)
   [#3007](https://github.com/realm/SwiftLint/issues/3007)
+* None.
 
 #### Bug Fixes